Single QgsBasicRelationWidget file
