+++ python/gui/auto_generated/qgsbasicrelationwidget.sip.in
@@ -103,6 +103,7 @@ Sets the title of the root groupbox
 };
 
 
+
 /************************************************************************
  * This file has been generated automatically from                      *
  *                                                                      *
